Protect Actions::task with its own Mutex with minimal scope
The overall idea is that we want to drop the global Streams Mutex or
limit its scope as much as possible.

This Mutex is protecting a lot of shared state including Wakers that
we need to poke to unblock waiting tasks when an event happened.

This particular change protects the Waker that is set by the future
driving the connection and makes all the rest of the code that consumes
it grab a Mutex to access it.

This is one of the simpler bits that we can separate as the use pattern
is pretty simple - the future driving the connection sets the Waker when
it has nothing else to do and all other code consumes it once state
changes and notifies the future driving the connection to wake up and do
some work.

This change by itself does not remove the global lock, since there is
still a lot of shared state that needs protection. On the flip side,
because the global lock is still there this change is safe
concurrency-wise.

Signed-off-by: Mikhail Krinkin <mkrinkin@microsoft.com>