Drop a SendBuffer lock completely
SendBuffer is just a wrapper aroud Buffer. Buffer is just a wrapper
around Slab, so basically SendBuffer and Buffer are just allocators.

This makes synchronizing those easier, since we can just replace the
Slab used underneath with a Slab from a sharded-slab crate that is
already thread safe.

A few caveats here that slightly complicate the matter:

1. Just because the Slab itself is thread safe it does not mean that the
   object we store there is safe to concurrently modify, so I had to use
   atomics for the next field in the Slot - we don't need any ordering
   guarantess, but using atomic makes the whole Slot Send, which is what
   we want.
2. There is no is_empty method for the Slab in sharded-slab crate and
   the only way to check if it is empty or not I could find is via an
   unique_iter method (by checking if it's an iterator over an empty
   sequence or not), but predictably unique_iter method requires a
   mutable self reference, so instead of using unique_iter method I
   added another atomic to count the number of items added to the buffer
   instead.

It's a bit dirty implementation because Slab in sharded-slab has a
different interface and exposes a few error conditions that could
potentially trigger that this change does not handle, but it should be
good enough for a PoC, to get us to a state where we could measure the
effects of all the changes.

Signed-off-by: Mikhail Krinkin <mkrinkin@microsoft.com>