Drop Deref traits for Ptr type
For reasons similar to Index and IndexMut traits Deref makes it a bit
harder to experiment because the return type of Deref must be a
reference.

I'm sure this will not be a long term problem in the future once I
settle on the final solution, because we could get a reference to a
Stream when we lock it and wouldn't need Ptr after that given that we
will have a stream reference (most likely).

For now I just replace implcit dereference with explicit methods.

Signed-off-by: Mikhail Krinkin <mkrinkin@microsoft.com>